GetKey Crash fixed, added condition to validate data field
