+++ .env
@@ -1,4 +1,4 @@
 GENERATE_SOURCEMAP=false
 SKIP_PREFLIGHT_CHECK=true
-REACT_APP_API_ENDPOINT = "https://console-stage.mobiledgex.net:443"
-REACT_APP_API_VM_ENDPOINT = "https://console-stage-vnc.mobiledgex.net:443"
+REACT_APP_API_ENDPOINT = "https://console-qa.mobiledgex.net:443"
+REACT_APP_API_VM_ENDPOINT = "https://console-qa-vnc.mobiledgex.net:443"
