+++ src/services/model/cloudlet.js
@@ -7,22 +7,25 @@ import { SHOW_CLOUDLET, SHOW_ORG_CLOUDLET, CREATE_CLOUDLET, STREAM_CLOUDLET, DEL
 const fields = formatter.fields;
 
 export const getKey = (data, isCreate) => {
-    let cloudlet = {}
-    cloudlet.key = {
-        organization: data[fields.operatorName],
-        name: data[fields.cloudletName]
+    if (data) {
+        let cloudlet = {}
+        cloudlet.key = {
+            organization: data[fields.operatorName],
+            name: data[fields.cloudletName]
+        }
+        if (isCreate) {
+            cloudlet.location = data[fields.cloudletLocation]
+            cloudlet.ip_support = data[fields.ipSupport]
+            cloudlet.num_dynamic_ips = parseInt(data[fields.numDynamicIPs])
+            cloudlet.physical_name = data[fields.physicalName]
+            cloudlet.platform_type = data[fields.platformType]
+        }
+        return ({
+            region: data[fields.region],
+            cloudlet: cloudlet
+        })
     }
-    if (isCreate) {
-        cloudlet.location = data[fields.cloudletLocation]
-        cloudlet.ip_support = data[fields.ipSupport]
-        cloudlet.num_dynamic_ips = parseInt(data[fields.numDynamicIPs])
-        cloudlet.physical_name = data[fields.physicalName]
-        cloudlet.platform_type = data[fields.platformType]
-    }
-    return ({
-        region: data[fields.region],
-        cloudlet: cloudlet
-    })
+    return {}
 }
 
 export const multiDataRequest = (keys, mcRequestList) => {
@@ -108,10 +111,6 @@ export const keys = [
     { field: fields.status, serverField: 'status', label: 'Status', dataType: TYPE_JSON },
     { field: fields.actions, label: 'Actions', sortable: false, visible: true, clickable: true }
 ]
-
-
-
-
 
 const customData = (value) => {
     value[fields.cloudletStatus] = 4
