+++ src/services/model/clusterInstance.js
@@ -77,38 +77,41 @@ export const showClusterInsts = (data) => {
 }
 
 export const clusterKey = (data, isCreate) => {
-    let clusterinst = {}
-    clusterinst.key = {
-        cluster_key: { name: data[fields.clusterName] },
-        cloudlet_key: { organization: data[fields.operatorName], name: data[fields.cloudletName] },
-        organization: data[fields.organizationName]
+    if (data) {
+        let clusterinst = {}
+        clusterinst.key = {
+            cluster_key: { name: data[fields.clusterName] },
+            cloudlet_key: { organization: data[fields.operatorName], name: data[fields.cloudletName] },
+            organization: data[fields.organizationName]
+        }
+        clusterinst.flavor = { name: data[fields.flavorName] }
+        if (isCreate) {
+            clusterinst.deployment = data[fields.deployment]
+            if (data[fields.ipAccess]) {
+                clusterinst.ip_access = parseInt(IPAccessLabel(data[fields.ipAccess]))
+            }
+            clusterinst.reservable = data[fields.reservable]
+            if (data[fields.reservedBy]) {
+                clusterinst.reserved_by = data[fields.reservedBy]
+            }
+            if (data[fields.numberOfMasters]) {
+                clusterinst.num_masters = parseInt(data[fields.numberOfMasters])
+            }
+            if (data[fields.num_nodes]) {
+                clusterinst.num_nodes = parseInt(data[fields.num_nodes])
+            }
+        }
+        return ({
+            region: data[fields.region],
+            clusterinst: clusterinst
+        })
     }
-    clusterinst.flavor = { name: data[fields.flavorName] }
-    if (isCreate) {
-        clusterinst.deployment = data[fields.deployment]
-        if (data[fields.ipAccess]) {
-            clusterinst.ip_access = parseInt(IPAccessLabel(data[fields.ipAccess]))
-        }
-        clusterinst.reservable = data[fields.reservable]
-        if (data[fields.reservedBy]) {
-            clusterinst.reserved_by = data[fields.reservedBy]
-        }
-        if (data[fields.numberOfMasters]) {
-            clusterinst.num_masters = parseInt(data[fields.numberOfMasters])
-        }
-        if (data[fields.num_nodes]) {
-            clusterinst.num_nodes = parseInt(data[fields.num_nodes])
-        }
-    }
-    return ({
-        region: data[fields.region],
-        clusterinst: clusterinst
-    })
+    return {}
 }
 
 export const getClusterInstList = async (self, data) => {
     return await serverData.showDataFromServer(self, showClusterInsts(data))
- }
+}
 
 export const createClusterInst = (data, callback) => {
     let requestData = clusterKey(data, true)
