+++ src/services/model/orgCloudlet.js
@@ -15,15 +15,18 @@ const keys = [
 ]
 
 export const getKey = (data) => {
-    return ({
-        region: data[fields.region],
-        cloudlet: {
-            key: {
-                organization: data[fields.operatorName],
-                name: data[fields.cloudletName]
+    if (data) {
+        return ({
+            region: data[fields.region],
+            cloudlet: {
+                key: {
+                    organization: data[fields.operatorName],
+                    name: data[fields.cloudletName]
+                }
             }
-        }
-    })
+        })
+    }
+    return {}
 }
 
 const customData = (value) => {
