+++ src/services/model/organization.js
@@ -9,24 +9,27 @@ export const keys = () => ([
     { field: fields.type, serverField: 'Type', label: 'Type', sortable: true, visible: true },
     { field: fields.phone, serverField: 'Phone', label: 'Phone', sortable: true, visible: true },
     { field: fields.address, serverField: 'Address', label: 'Address', sortable: true, visible: true },
-    { field: 'manage', label: 'Manage', visible: false, clickable: true},
+    { field: 'manage', label: 'Manage', visible: false, clickable: true },
     { field: fields.actions, label: 'Actions', visible: true, clickable: true }
 ])
 
 export const getKey = (data) => {
-    return ({
-        name: data[fields.organizationName],
-        type: data[fields.type],
-        address: data[fields.address],
-        phone: data[fields.phone]
-    })
+    if (data) {
+        return ({
+            name: data[fields.organizationName],
+            type: data[fields.type],
+            address: data[fields.address],
+            phone: data[fields.phone]
+        })
+    }
+    return {}
 }
 
 export const showOrganizations = (data) => {
     return { method: SHOW_ORG, data: data }
 }
 
-export const getOrganizationList = async(self, data) => {
+export const getOrganizationList = async (self, data) => {
     let dataList = []
     if (formatter.getOrganization()) {
         let organization = {}
