+++ src/services/model/users.js
@@ -1,6 +1,6 @@
 import { fields, formatData } from './format'
 import * as serverData from './serverData'
-import { SHOW_USERS, DELETE_USER, ADD_USER_ROLE} from './endPointTypes'
+import { SHOW_USERS, DELETE_USER, ADD_USER_ROLE } from './endPointTypes'
 
 
 
@@ -12,11 +12,14 @@ export const keys = [
 ]
 
 export const getKey = (data) => {
-    return ({
-        org: data[fields.organizationName],
-        username: data[fields.username],
-        role: data[fields.role]
-    })
+    if (data) {
+        return ({
+            org: data[fields.organizationName],
+            username: data[fields.username],
+            role: data[fields.role]
+        })
+    }
+    return {}
 }
 
 export const showUsers = (data) => {
